refactor(functions): address review comments for declarative security API check
- Move checkDeclarativeSecurityApisEnabled to src/deploy/functions/ensure.ts
- Scope fail-open error handling specifically to HTTP 403 and PERMISSION_DENIED
- Improve debug log message phrasing and preserve Winston error stack trace
- Use Sinon .withArgs(...) in prepare.spec.ts and shorten expectation assertions
- Add unit test verifying non-permission errors are re-thrown
